vMul option and set default option function
Added a new option called vMul (value multiplier). The set function
divides the value with vMul and the get function multiplies the value
with vMul. We needed this functionality because we save the values in
1/100 of the currency in our database.
moved the "default" object out of the init function to make it possible
to implement $.autoNumeric.defaults function which makes it possible to
set the default option object.