updates to the init, set, get and getString methods
Fixed the "getSting" method - it use to returned an error when no values
were entered
Modified the "init" method to better handle default and pre-existing
values
Modified the "set" method - removed the routine that checked for values
less than .000001 and greater than -1 and placed it in a separate
function named checkValue()
Modified the "get" method:
Added a call to the checkValue() function - this corrects returned
values. Example - when the input value was "12." the returned value was
"12." - it now returns "12"
When no numeric character is present the returned value is an empty
string "".