A potential fix for when parseStr is inadvertantly passed an integer.
It's interesting - the function is called "parseStr" yet the first conditional statement does an explicit type check for integers. Either way, I'm guessing that this was meant to catch "all cases of an integer" as the following conditions rely on the n parameter being a string. This fix casts the int to a string if not already at this point.