Simplify the `init` function so that it's easier to read and understand (from 406 to 33 lines!).
Extract every event listener to its own function.
Extract the test to see if the tag and type on which autoNumeric is initialized is supported.
Extract the initial formatting on load routine.
Extract the initial settings generation based on what the user set in the html or via the options.
Extract `isValidPasteText()` and `preparePastedText()` function out of the `paste` event handler.
Add a default value for the `update` parameter of the `getHolder()` function.