Add a new `failOnUnknownOption` option.
This option is the 'strict mode' (aka 'debug' mode), which allows autoNumeric to strictly analyse the options passed, and fails if an unknown options is used in the settings object.
By default, this is set to `false`.
You should set it to `true` if you want to make sure you are only using 'pure' autoNumeric settings objects in your code.