Implement feature request #183 Manage invalid results when trying to paste any number
Add a new `onInvalidPaste` option that alters the behavior of autoNumeric when pasting a number results in an invalid value.
Upgrade the `setElementSelection()` function so that it can accept only one caret position.
Create 11 helper functions specially designed for managing the paste event.
Rename `autoStrip()` to `stripAllNonNumberCharacters()`.
Format the file header.
Fix `correctMDecOption()` so that the original setting `decimalPlacesOverride` is always saved.