Fix issue #187 Add a format() callback
A 'autoNumeric:formatted' event is sent as soon as the input value is formatted.
This happens after each key press, if the resulting value has changed, which means if the user hit a key that does not result in the input value changing, no such event is sent.
If you need to modify the input value somehow on blur, you can listen to the 'change' event that is now correctly sent.