Simplify the event keyCode management and fix issue #318.
Fix issue #318 `this.selection` can be uninitialized if you focus on an input via the `Tab` key.
Add the `keyName` object that list the key values as defined is the KeyboardEvent Key_Values.
Rename the `key()` function to `keyCodeNumber()`.
Split `_updateFieldProperties()` into `_updateAutoNumericHolderProperties()` and `_updateAutoNumericHolderEventKeycode()`.
`_updateAutoNumericHolderProperties()` only update the value and selection in each event handler, and resets the 'processed' and 'formatted' states.
`_updateAutoNumericHolderEventKeycode()` is called only once on `keypress`, and set the event keyCode into the AutoNumericHolder object.
Remove the need to save `this.ctrlKey`, `this.cmdKey` and `this.shiftKey` three times on each key stroke.
Rename `this.kdCode` into `this.eventKeyCode`, and use that variable globally in the AutoNumericHolder object.
Fix `_normalizeParts()` so that a '0' entered via the numpad is managed as well.
Complete the substitution of magic numbers with `keyCode` values.
Modify `_processCharacterInsertion()` so that it take the event as an argument, and therefore can directly use `e.key`.
Simplify `_formatValue()` tests.