Fix issue #346 Add a `showPositiveSign` option to display the positive sign wherever needed
Upgrade `isNegative()` so that it can search for a negative sign anywhere in the value.
Add `isZeroOrHasNoValue()` to check if a numeric string represent a value of zero, or is just empty.
Fix `correctNegativePositiveSignPlacementOption()` so that it does not overwrite the `negativePositiveSignPlacement` option if it has been set by the user.
Removes an unnecessary ternary operator.
Fix the bug when trying to unformat a negative value that has a the negative sign not positioned on the far left.
Create a `isNegativeStrict()` function that test if a numeric string has a negative sign on its very first character.
Use `isNegative()` and `isNegativeStrict()` everywhere where the same test was made.