Fix issue #310 Setup Webdriver.io for end-to-end (e2e) testing
Fix the test for a negative sign in `_getUnformattedLeftAndRightPartAroundTheSelection()` and `_normalizeParts()`.
Create a `character()` function that returns the real character being typed in events.
This `character()` function solves the potential compatibility problem with old browsers that do not correctly implement the `event.key` attribute.
Replace all calls to `event.key` with a call to the new `character()` function.
This changes allow autoNumeric to work with the Selenium server (which does not set correctly the `e.key` attribute, cf. related issue https://github.com/webdriverio/webdriverio/issues/1829).
Add the `test/e2e/index.html` file that store all the form inputs and autoNumeric configurations for the end-to-end tests.