Fix issue #387 Add a 'cancellable' feature Fix issue #391 The currency symbol is selected when focusing on an input via the `Tab` key, when `selectNumberOnly` is set to `true`
Refactor the code to create a `_selectOnlyNumbers()` function that extract that behavior for re-use.
Create a `_select()` function that select the whole element content, while respecting the `selectNumberOnly` option.
Create a `_defaultSelectAll()` function that select the whole element content, including all characters.
Modify the `setElementSelection()` calls to simplify them with the ability to use one argument instead of two when the `start` and `end` position are the same.
Add a feature where when the user hit 'Escape', the element content is selected (cf. issue #387).