Fix issue #396 Allow autoNumeric to manage user inputs in `contentEditable`-enabled elements
Simplify the `_formatDefaultValueOnPageLoad()` method signature.
Simplify how the `update()` function retrieve the current raw value and one of its test.
Remove any direct access to `this.domElement.value`, and use the `getElementValue()` to access the element `value` or `textContent`, and `setElementValue()` to set it.
Rewrite the `getElementSelection()` and `setElementSelection()` functions in order to manage non-input DOM elements.
Strengthen `getElementValue()` when managing non-input DOM elements.

Signed-off-by: Alexandre Bonneau <alexandre.bonneau@linuxfr.eu>