Finish merging the fixes for #403 and #401 into `next`, and fix the introduced bugs
Make sure that if an element is focused, and a `mouseleave` event is captured, we do not unformat the element value (with the scale options).
Make sure if the element value is empty, that the scaleDivisor option do not convert it to `'0.00'` on `mouseenter`.
Remove a ternary operator which lead to a variable overwriting itself.
Fix the end-to-end tests which test the percentage with the scale options.

Signed-off-by: Alexandre Bonneau <alexandre.bonneau@linuxfr.eu>