Add unit tests for the new `options.*` methods
Fix the issue where having `emptyInputBehavior` equal to `always` would not reformat the AutoNumeric element correctly if its value is set to the empty string `''`.
Fix the issue where having `emptyInputBehavior` equal to `zero` would not reformat the AutoNumeric element correctly if its value is set to the empty string `''`.
Fix the `_mergeCurrencySignNegativePositiveSignAndValue()` function that did not manage all `negativePositiveSignPlacement` combinations.
Simplify the `_mergeCurrencySignNegativePositiveSignAndValue()` function.
Add function chaining to the `options.*` methods.

Signed-off-by: Alexandre Bonneau <alexandre.bonneau@linuxfr.eu>