Fix issue #427 `autoUnformat()` converts `digitGroupSeparator` set to `.` as the decimal character
Modify `AutoNumeric.unformat()` so that 'real' javascript number are always directly returned, without taking into account the options passed (as it was supposed to do previously).
Correctly take into account the number of decimal places, the negative brackets, rounding and the suffix text options when unformatting with `AutoNumeric.unformat()`.
Complete the `AutoNumeric.unformat()` unit tests.
Modify `_removeBrackets()` so that we can only remove the brackets, without reordering the negative sign, currency symbol and value according to the settings.

Signed-off-by: Alexandre Bonneau <alexandre.bonneau@linuxfr.eu>