Fix the `'autoNumeric:formatted'` event not being correctly sent if the AutoNumeric element was formatted without a `keyup` event.
The event is now correctly sent when the value is set to empty, or when using `unformat()`, `unformatLocalized()` and `wipe()`, as well as when the user uses the wheel event to change the element value, or the `alt + mouse hover` feature, or just hover the element that has a `negativeBracketsTypeOnBlur` option set, or on the initial format on load.
Fix the fact that `'autoNumeric:formatted'` was not sent when pasting valid values.
Gather the AutoNumeric event names in a single configuration variable `AutoNumeric.events`.

Signed-off-by: Alexandre Bonneau <alexandre.bonneau@linuxfr.eu>