Issue #709 Behaviour with '+' and '-' keypress
This changes the behaviour of the '+' and '-' keys.
Before, each of them toggled the sign from positive to negative and back.
Which means, in an empty field, that you get a value of "-123" when typing the keys "+123". And hitting the '+' key several times made the number positive, negative, positive again, and so on.
After the change, the '+' key makes a negative number positive, and has no effect on a positive number.
And the '-' key makes a positive number negative, and has no effect on a negative number.