Fixes the caret positioning in some specific cases
When entering a decimal character (i.e. '.') while the caret is next to the existing decimal character, the caret was jumping to the end of the fields. This fixes this annoying behavior, by allowing the caret to be move one place to the right, as before.
Updates the jsdoc.
Fixes some end-to-end tests (this is a WIP).

Signed-off-by: Alexandre Bonneau <alexandre.bonneau@linuxfr.eu>